[release] fix dataproc failures (#15199)
Fixes the following dataproc test failures in the lastest release
commit:

- VEP serialisation failure
- Warnings from mismatched spark version
- Java/Scala library version conflicts
- Logging configuration errors
- pyarrow abi conflicts in dataproc 2.2.x
- NullPointerException from premature hadoop.fs.closeall call

Major changes:

1. dataproc 2.3.17-debian12

pyarrow had api conflicts with numpy2 in dataproc 2.2.x. dataproc 2.3.x
fixes this
and also provides many of the python libraries we install in
`init_notebook.py`.

1. build.mill

Spark provides many of our runtime dependencies and places its versions
of libraries
ahead of `hail-all-spark.jar` on the classpath. Unless we relocate these
libraries,
we need explicit compile-only dependences on the libraries that we use
directly but
spark ships. Turns out many of spark 3.5.2's 3p libraries are the ones
we need so
don't have to relocate as many!

1. VEP

To reduce what we need to serialise and work around VEP serialisation
failures brought
about by the log4j-api-scala introduction, VEP implementation is now
anonymous. There
are no other functional changes therein.

This change has low security impact on the Broad-managed hail batch
service in GCP,
with changes limited to the version of spark and java logging library
used for query-on-batch.