[release] fix classpath error for qob (#15201)
#15199 fixed running in dataproc. It broke query-on-batch, however, as
the hail-all-spark.jar assembly bundled in different versions of
libraries that spark provides.

To fix this, I've explicitly excluded many of the libraries spark
provides from the assembly. This is by no means complete because finding
which library causes which dependency to get bundled is not trivial.
Mill's dependence resolving isn't the best.

To make the QoB classpath behave like dataproc, I've positioned the
spark libraries before the hail jar on the worker's classpath. I believe
this should not cause issues for previous hail versions as they've
worked on dataproc where this is the case.

This change has low impact on the Broad-managed hail batch deployment in
GCP.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec