[auth] Add 'inactive' status to the list of account errors (#15210)
A user reported a 500 when trying to log in on zulip ([#Hail Batch
support > unable to log in -- 500
error](https://hail.zulipchat.com/#narrow/channel/223457-Hail-Batch-support/topic/unable.20to.20log.20in.20--.20500.20error/with/563247607)).

We run into this internal server error because we were failing
[this](https://github.com/hail-is/hail/blob/5d54486fcac06f6b7b2e8af380f812d1552343fc/auth/auth/auth.py#L438)
assertion, as were not handling the 'inactive' status. This change adds
the 'inactive' status to the list of errors, and also makes sure to set
an http error status when rendering the `account-errors.html` template.

## Security Assessment
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating
- This change has low security impact

### Impact Description
Change the response codes to one portion of the login flow, and make it
so that the server doesn't fail when handling login of an inactive user.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec